Refactor crop yield chart and iqr/median chart components
